Move adding bounds to member
